break stripe seats dependency loop
